requirements.txt
pycryptodomex==3.9.4
requests==2.22.0
websocket-client==0.57.0